backend/requirements.txt
fastapi
spotipy
openai
anthropic
rapidfuzz
uvicorn
httpx
python-dotenv
mangum
